Re-enables findBugs; fixes found bugs
