+++ core/src/main/java/org/cirdles/topsoil/plot/JavaScriptBridge.java
@@ -1,0 +1,14 @@
+package org.cirdles.topsoil.plot;
+
+/**
+ * Intended to be added as a member to a JavaScriptPlot, so that the JavaScript may make upcalls to JavaFX.
+ *
+ * @author Emily Coleman
+ */
+public class JavaScriptBridge {
+
+    public void println(String s) {
+            System.out.println(s);
+        }
+
+}
